examples with mysql instead of postgres
