+++ .env
@@ -22,5 +22,5 @@ APP_SECRET=7ec00b95a6f0c871e20334a014dd7c04
 ###> doctrine/doctrine-bundle ###
 # Format described at https://www.doctrine-project.org/projects/doctrine-dbal/en/latest/reference/configuration.html#connecting-using-a-url
 # IMPORTANT: You MUST configure your server version, either here or in config/packages/doctrine.yaml
-DATABASE_URL="postgresql://app:!ChangeMe!@database:5432/app?serverVersion=17&charset=utf8"
+DATABASE_URL="mysql://root:!ChangeMe!@database:3306/app?serverVersion=9&charset=utf8"
 ###< doctrine/doctrine-bundle ###
