+++ docker/Dockerfile
@@ -14,7 +14,7 @@ RUN apk -U upgrade && apk add --no-cache \
     php83-mbstring \
     php83-openssl \
     php83-pcntl \
-    php83-pdo_pgsql \
+    php83-pdo_mysql \
     php83-phar \
     php83-session \
     php83-tokenizer \
