--- src/Repository/AuctionJsonbRepository.php
@@ -1,83 +1,0 @@
-<?php
-
-namespace App\Repository;
-
-use App\Entity\AuctionJsonb;
-use App\Model\Item;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\DBAL\Statement;
-use Doctrine\Persistence\ManagerRegistry;
-use Symfony\Component\Serializer\SerializerInterface;
-
-/**
- * @extends ServiceEntityRepository<AuctionJsonb>
- */
-class AuctionJsonbRepository extends ServiceEntityRepository implements AuctionInterface
-{
-    public function __construct(
-        ManagerRegistry $registry,
-        private readonly SerializerInterface $serializer,
-        private readonly ItemExtractor $itemExtractor,
-    ) {
-        parent::__construct($registry, AuctionJsonb::class);
-    }
-
-    public function truncate(): void
-    {
-        $cmd = $this->getClassMetadata();
-        $connection = $this->getEntityManager()->getConnection();
-        $dbPlatform = $connection->getDatabasePlatform();
-        $q = $dbPlatform->getTruncateTableSql($cmd->getTableName());
-        $connection->executeStatement($q);
-    }
-
-    public function createAuction(Item $item): AuctionJsonb
-    {
-        $data = $this->serializer->normalize($item);
-        $auction = new AuctionJsonb($item->title, $item->startDate, $item->endDate, $data);
-        $this->getEntityManager()->persist($auction);
-
-        return $auction;
-    }
-
-    public function getInsertStatement(int $batchSize): Statement
-    {
-        $statement = 'INSERT INTO '.$this->getClassMetadata()->getTableName().' (title, start_date, end_date, item) VALUES ';
-        $values = [];
-        for ($i = 0; $i<$batchSize; $i++) {
-            $values[] = "(:title$i, :startDate$i, :endDate$i, :item$i)";
-        }
-        $values = implode(',', $values);
-
-        return $this->getEntityManager()
-            ->getConnection()
-            ->prepare($statement.$values)
-        ;
-    }
-
-    public function updateAuction(AuctionJsonb $auction, Item $item): void
-    {
-        $data = $this->serializer->normalize($item);
-        $auction->setTitle($item->title);
-        $auction->setStartDate($item->startDate);
-        $auction->setEndDate($item->endDate);
-        $auction->setItem($data);
-    }
-
-    public function extractItem(AuctionJsonb $auction): Item
-    {
-        return $this->itemExtractor->extractItem($auction->getItem());
-    }
-
-    public function countAuthor(string $authorName): int
-    {
-        $qb = $this->createQueryBuilder('a');
-        $qb->select('COUNT(a)')
-            // Need to qualify with a. in front of item
-            ->where("JSON_GET_TEXT(a.item, 'author') = :author")
-            ->setParameter('author', $authorName)
-        ;
-
-        return $qb->getQuery()->getSingleScalarResult();
-    }
-}
